Fix the types for the gradient
Move back to memoryviews and use the new 'get buffer type'
method from the cy_utils.pxd

+++ menpo/feature/cpp/central_difference.h
@@ -1,24 +1,24 @@
 #include <stdio.h>
 
-static inline long long SUB2IND(const long long j, const long long i, const long long k,
-                                const long long row_size, const long long col_size,
-                                const long long n_channels) {
+static inline size_t SUB2IND(const size_t j, const size_t i, const size_t k,
+                             const size_t row_size, const size_t col_size,
+                             const size_t n_channels) {
     //return ((i + col_size * j) * n_channels) + k;
     return (i + col_size * j) + (row_size * col_size * k);
 }
 
 template<typename T>
-void central_difference(const T* in, const long long rows,
-                        const long long cols, const long long n_channels,
+void central_difference(const T* in, const size_t rows, const size_t cols,
+                        const size_t n_channels,
                         T* out) {
-    const long long n_output_channels = n_channels * 2;
-    long long output_index = 0;
+    const size_t n_output_channels = n_channels * 2;
+    size_t output_index = 0;
 
     #pragma omp for
-    for (long long k = 0; k < n_channels; ++k) {
+    for (size_t k = 0; k < n_channels; ++k) {
         // row-derivative
-        for (long long i = 0; i < cols; ++i) {
-            for (long long j = 0; j < rows; ++j) {
+        for (size_t i = 0; i < cols; ++i) {
+            for (size_t j = 0; j < rows; ++j) {
                 if (j == 0) {
                     output_index = SUB2IND(0, i, k, rows, cols, n_output_channels);
                     out[output_index] = in[SUB2IND(1, i, k, rows, cols, n_channels)] - in[SUB2IND(0, i, k, rows, cols, n_channels)];
@@ -35,8 +35,8 @@ void central_difference(const T* in, const long long rows,
         }
 
         // column-derivative
-        for (long long j = 0; j < rows; ++j) {
-            for (long long i = 0; i < cols; ++i) {
+        for (size_t j = 0; j < rows; ++j) {
+            for (size_t i = 0; i < cols; ++i) {
                 if (i == 0) {
                     output_index = SUB2IND(j, 0, n_channels + k , rows, cols, n_output_channels);
                     out[output_index] = in[SUB2IND(j, 1, k, rows, cols, n_channels)] - in[SUB2IND(j, 0, k, rows, cols, n_channels)];
